requirements.txt
ping3
psutil
requests
dnspython
paramiko
scapy
